update: add jumps from swap txn flow to recv txn flow and back for BTC
For the recipient address derivation during the swap txn flow, we jump from the swap txn flow state to the recv address derivation state. At the last state of the recv address derivation based on a flag we either return back to the swap txn flow or transmit the data to the app as usual.