fix (build): Reduce time to generate git commit-id
The current call to execute_process uses `bash` to execute the test command
for checking dirtiness of the git worktree. It takes a long time about a
few seconds when executed on windows in git-bash. Replacing with `sh` changes
the behaviour and it takes under a second (as expected). This is mostly
observed on systems where WSL is enabled in which case, running an instance
of bash takes longer.